fastapi-backend/requirements.txt
# NEXUS Backend Dependencies
# FastAPI and Server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Authentication
pyjwt==2.8.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

# Validation and Settings
pydantic==2.5.2
pydantic-settings==2.1.0
email-validator==2.1.0

# Async Support
httpx==0.25.2
websockets==12.0

# Data Processing
numpy==1.26.2

# Logging
loguru==0.7.2

# Environment
python-dotenv==1.0.0